Fixes MyPy errors and runs black

+++ csv_fun/parse_results.py
@@ -6,4 +6,4 @@ from .dtos.narababy_event_row import NarababyEventRow
 class ParseResults:
     data: list[NarababyEventRow]
     rows_processed: int
-    time_elapsed: int
+    time_elapsed: float
